% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otBasicStats.R
\name{ot_basic_stats_vec}
\alias{ot_basic_stats_vec}
\title{Computing various statistics}
\usage{
ot_basic_stats_vec(x)
}
\arguments{
\item{x}{a variable to summarise}
}
\description{
Computing various statistics
}
