% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_anova}
\alias{ot_print_anova}
\title{Print "otANOVA" resutls}
\usage{
ot_print_anova(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object make by ot_functions.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
otANOVA is a function to print out "otANOVA" results.
}
