% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_basic_stats}
\alias{ot_print_basic_stats}
\title{Print "otBasicStats" results}
\usage{
ot_print_basic_stats(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object computed by "otBasicStats"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print "otBasicStats" results
}
