% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_colleration}
\alias{ot_print_colleration}
\title{Print out function for "otCorrelation"}
\usage{
ot_print_colleration(otm_obj, ...)
}
\arguments{
\item{otm_obj}{any otmR object computed by "otCorrelation"}
}
\description{
Print out function for "otCorrelation"
}
