% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_crosstable}
\alias{ot_print_crosstable}
\title{Print "otCrossTable" results}
\usage{
ot_print_crosstable(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object computed by "otCrossTable"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print "otCrossTable" results
}
