% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_frequencytable}
\alias{ot_print_frequencytable}
\title{Print "otFrequencyTable" results}
\usage{
ot_print_frequencytable(otm_obj)
}
\arguments{
\item{otm_obj}{an object computed by "otFrequencyTable"}
}
\description{
Print "otFrequencyTable" results
}
