% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_logistic_regression}
\alias{ot_print_logistic_regression}
\title{Print "otLogisticRegression" results}
\usage{
ot_print_logistic_regression(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object computed by "otLogisticRegression"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print "otLogisticRegression" results
}
