% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_ttest}
\alias{ot_print_ttest}
\title{Print "otTTest" results}
\usage{
ot_print_ttest(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object computed by "otTTest"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print "otTTest" results
}
