% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ot_set_variable.R
\name{ot_set_variable}
\alias{ot_set_variable}
\title{Setting value labels to a factor variable}
\usage{
ot_set_variable(x, v_label = NULL)
}
\arguments{
\item{x}{a vector to set value labels.}

\item{v_label}{a label string vector: the format "level=label".}
}
\description{
ot_set_variable is the function to put value labels like SPSS
to a integer variable.  It returns a factor variable.
}
