% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otANOVA.R
\name{otAnova}
\alias{otAnova}
\title{Anova Tables}
\usage{
otAnova(data, model = NULL)
}
\arguments{
\item{data}{a data.frame to compute ANOVA tables.}

\item{model}{a formula to compute ANOVA tables.}
}
\description{
otAnova is a funtion to compute ANOVA tables.
}
