% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otBasicStats.R
\name{otBasicStats}
\alias{otBasicStats}
\title{Calculating Basic Statistics for ratio scale variables}
\usage{
otBasicStats(data)
}
\arguments{
\item{data}{target data.frame}
}
\description{
Calculating Basic Statistics for ratio scale variables
}
