% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otCorrelation.R
\name{otCorrelation}
\alias{otCorrelation}
\title{Computing correlation matrix from data.frame}
\usage{
otCorrelation(
  data,
  use = "pairwise.complete.obs",
  method = c("pearson", "kendall", "spearman")
)
}
\arguments{
\item{data}{a data.frame that contains numeric variables only}

\item{use}{an optional character string which is one of the strings
"everything", "all.obs", "complete.obs", na.or.complete", or
"pairwise.complete.obs".  See the heflp "cor".}

\item{method}{a character string indicating which correlation coefficient is
to be computed.  One of "pearson", "kendall", or "spearman" can be abbreviated.}
}
\description{
Computing correlation matrix from data.frame
}
