% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otCrossTable.R
\name{otCrossTable}
\alias{otCrossTable}
\title{Cross Table}
\usage{
otCrossTable(data, model = NULL, omit.na = TRUE, is.test = TRUE)
}
\arguments{
\item{data}{a data.frame object including two variables}

\item{model}{a formula "var_a ~ var_b" style}

\item{omit.na}{a logical value.  If TRUE, NA is deleted
from a cross table.}

\item{is.test}{a logical value.  If TRUE, chisq.test and
fisher.test are submitted.}
}
\description{
otCrossTable is a function to make 2-dimensional cross table
from two variables.  It is similar to the function "table", but
otCrossTable use a formula to set two variables.  This is why
other ot-functions use formula style.
}
