% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otFrequencyTable.R
\name{otFrequencyTable}
\alias{otFrequencyTable}
\title{Frequency Table}
\usage{
otFrequencyTable(data, omit.na = TRUE)
}
\arguments{
\item{data}{a data.set object computing frequency tables}

\item{omit.na}{a logical indicating whether NA is to be omitted or not}
}
\description{
otFrequencyTable is to make frequency tables of some variables.
}
