% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otGlm.R
\name{otGlm}
\alias{otGlm}
\title{Fitting Generalized Linear Models for otmR}
\usage{
otGlm(data, model = NULL, is.residual = FALSE)
}
\arguments{
\item{data}{a data.frame object including both a dependent variables and independent
variables.}

\item{model}{an object of class "formula": a symbolic description
of the model to be fitted.}
}
\description{
otGlm is used to fit generalized linear models.  otGlm returns
a data.frame including simple results of stats::glm().
}
