% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otLogisticRegression.R
\name{otLogisticRegression}
\alias{otLogisticRegression}
\title{Logistic Regression}
\usage{
otLogisticRegression(data, model = NULL, is.residual = FALSE)
}
\arguments{
\item{data}{a data.frame object including both a dependent variables and independent
variables.}

\item{model}{an object of class "formula": a symbolic description
of the model to be fitted.}

\item{is.residual}{a logical value indicating whether return object
have residual information or not.}
}
\description{
otLogisticRegression is used to fit logistic models,
using glm functions.  An independent variable must have
0-1 value.
}
