% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otReadData.R
\name{otReadData}
\alias{otReadData}
\title{otReadData}
\usage{
otReadData(path, sheet = NULL, ...)
}
\arguments{
\item{path}{path to the xls/xlsx file.}

\item{sheet}{sheet to read.  Ether a string (the name of sheet),
or an integer (the position of the sheet).}

\item{...}{further arguments passed to or from other methods.}
}
\description{
otReadData is a function to make a data.frame from a xlsx book file.
If a sheet2 on a xlsx book file includes descriptions of each variables,
this function mutates a numeric vector to a factor one.
}
