% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otTTest.R
\name{otTTest}
\alias{otTTest}
\title{Student's t-Test}
\usage{
otTTest(data, model, p.value = 0.05)
}
\arguments{
\item{data}{a data.frame object including including both a
dependent variables and independent
variables.}

\item{model}{a formula of the same form lhs ~ rhs as stats::t.test}

\item{p.value}{a probability of significant level}
}
\description{
otTTest is a function of statistical test of means for two sample.
}
