% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_basic_stats}
\alias{ot_print_basic_stats}
\title{Print out function "otBasicStatis"}
\usage{
ot_print_basic_stats(otm_obj, ...)
}
\arguments{
\item{otm_obj}{any otmR object}
}
\description{
Print out function "otBasicStatis"
}
