% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{ot_print_colleration}
\alias{ot_print_colleration}
\title{Print "otCorrelation" results}
\usage{
ot_print_colleration(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object computed by "otCorrelation"}
}
\description{
Print "otCorrelation" results
}
