% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint 2.R, R/otPrint.R
\name{otPrint}
\alias{otPrint}
\title{Print out function for otmR}
\usage{
otPrint(otm_obj, ...)

otPrint(otm_obj, ...)
}
\arguments{
\item{otm_obj}{An object made by ot_functions}
}
\description{
Print out function for otmR

Print out function for otmR
}
