% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otPrint.R
\name{otPrint}
\alias{otPrint}
\title{Print out function for otmR}
\usage{
otPrint(otm_obj, ...)
}
\arguments{
\item{otm_obj}{an object made by ot_functions.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print out function for otmR
}
